test: update and move tests for count() method for multiple primary keys
