+++ test/functional/query-builder/count/entity/Test.ts
@@ -1,4 +1,4 @@
-import {Entity, PrimaryColumn} from "../../../../src";
+import {Entity, PrimaryColumn} from "../../../../../src";
 
 @Entity("tests")
 export class Test {
