+++ test/functional/query-builder/count/query-builder-count.ts
@@ -1,16 +1,15 @@
-import {closeTestingConnections, createTestingConnections, reloadTestingDatabases} from "../../utils/test-utils";
-import {Connection} from "../../../src/connection/Connection";
+import {closeTestingConnections, createTestingConnections, reloadTestingDatabases} from "../../../utils/test-utils";
+import {Connection} from "../../../../src/connection/Connection";
 import {expect} from "chai";
 import {Test} from "./entity/Test";
 
-describe("other issues > cockroachdb count query for multiple primary keys", () => {
+describe("query builder > count", () => {
 
     let connections: Connection[];
     before(async () => connections = await createTestingConnections({
         entities: [Test],
         schemaCreate: true,
         dropSchema: true,
-        enabledDrivers: ["cockroachdb"]
     }));
     beforeEach(() => reloadTestingDatabases(connections));
     after(() => closeTestingConnections(connections));
